Add a README
Git's home page is located at https://git-scm.com/. Two subdomains exist for
this domain: `www` and `book`. The `www` one is simply referring to the same
site as `git-scm.com`, and `book` (according to archive.org) should do the
same, but currently does not.

To remedy this, I am testing whether adding another GitHub Pages site
(to be reachable via https://book.git-scm.com/ and simply redirecting to
the actual ProGit book) does the job.

Let's describe the idea in the README.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>